SFML

+++ Controler/jeu.h
@@ -2,6 +2,7 @@
 #define JEU_H_INCLUDED
 
 #include "../Model/Plateau.h"
+#include <SFML/Graphics.hpp>
 
 
 
@@ -10,7 +11,7 @@ void openAllCase(Plateau* p);
 void contamination(Plateau* p, int pos_x, int pos_y);
 void flag(Plateau* p, int pos_x, int pos_y);
 
-void lancerPartie(int size_x, int size_y, int mine);
+void lancerPartie(int size_x, int size_y, float mine);
 int isWin(Plateau* p);
